naming refactor for locale files
