+++ packages/react-ui/src/localization/es.js
